Updates to the MQTT classes and general cleanup
Made several adjustments to the MQTTService class:
- Fixed to only look at the baseTopic/listenTopic MQTT topic for
incoming commands.  This leaves a lot less room for doing something
unexpected.
- Added a baseTopic/listenTopic/receipt MQTT topic that includes the
state of the most recent received incoming command.  Really, just if the
command was valid or not... it does not reflect if the Panel actually
actioned the command

Updated the AlarmTrouble class so that the myAddress function did not
return null.  This resulted in a [baseTopic]null MQTT topic which wasn't
desired.  All events are now in the "event" subtopic

Updated the NotificationPacket class to have a valid "toString"
function.  This allows easier debugging in the logs as the topic,
message and action are printed instead of the object ID

Added an updated jar with the changes.